Check the transition probabilities for numeric values and unity row sum

Check the transition probabilities for numeric values and unity row sum

check_trans_prob(trans_mat)

Arguments

trans_mat
transition matrix

Value

0 if they add to 1 else error

Details

checking for rowsum - checks for the class of transition matrix, value of rowsum (to be 1) and numeric values

Examples

tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm <- populate_transition_matrix(2, tmat, list_prob = c(0.5, 0.5, 0, 1))

#> [1] "The transition matrix as explained"
#>    transition number        probability name from from state to to state
#> 1:                 1 prob_Healthy_to_Healthy    1    Healthy  1  Healthy
#> 2:                 2    prob_Healthy_to_Dead    1    Healthy  2     Dead
#> 3:                 3    prob_Dead_to_Healthy    2       Dead  1  Healthy
#> 4:                 4       prob_Dead_to_Dead    2       Dead  2     Dead

check_trans_prob(tm)

#> [1] 0