Function to extract the unit hospital inpatient admission by matching HRG code

Function to extract the unit hospital inpatient admission by matching HRG code

get_cost_inpatient_hrg(
  hrg,
  ref_cost_data_file,
  col_name_hrg_code,
  unit_cost_col,
  sheet = NULL
)

Arguments

hrg
the form of medication either tablet or patch

ref_cost_data_file
file that has unit cost

col_name_hrg_code
name of the column that has the hrg code

unit_cost_col
name of the colum with the unit cost

sheet
sheet if excel file is given

Value

unit cost the unit cost matching the hrg code

Examples

ref_cost_data_file <- system.file("extdata",
"National_schedule_of_NHS_costs_2019.csv", package = "packDAMipd")
get_cost_inpatient_hrg("AA22C", ref_cost_data_file, "Currency_Code",
"National_Average_Unit_Cost")

#> [1] 5053