Get the parameter values using the provided statistical regression methods

Get the parameter values using the provided statistical regression methods

get_parameter_estimated_regression(
  param_to_be_estimated,
  data,
  method,
  indep_var,
  info_get_method = NA,
  info_distribution = NA,
  covariates = NA,
  timevar_survival = NA,
  interaction = FALSE,
  fix_eff = NA,
  fix_eff_interact_vars = NA,
  random_intercept_vars = NA,
  nested_intercept_vars_pairs = NA,
  cross_intercept_vars_pairs = NA,
  uncorrel_slope_intercept_pairs = NA,
  random_slope_intercept_pairs = NA,
  naaction = "stats::na.omit",
  param2_to_be_estimated = NA,
  covariates2 = NA,
  interaction2 = FALSE,
  link = NA,
  cluster_var = NA,
  package_mixed_model = NA
)

Arguments

param_to_be_estimated
parameter of interest

data
data to be provided or the data file containing dataset

method
method of estimation (for example, linear, logistic regression etc)

indep_var
the independent variable (column name in data file)

info_get_method
additional information on methods e.g Kaplan-Meier ot hazard

info_distribution
distribution name eg. for logistic regression -binomial

covariates
list of covariates-calculations to be done before passing

timevar_survival
time variable for survival analysis

interaction
boolean value to indicate interaction in the case of linear regression

fix_eff
boolean value to indicate interaction in the case of linear regression

fix_eff_interact_vars
boolean value to indicate interaction in the case of linear regression

random_intercept_vars
boolean value to indicate interaction in the case of linear regression

nested_intercept_vars_pairs
boolean value to indicate interaction in the case of linear regression

cross_intercept_vars_pairs
boolean value to indicate interaction in the case of linear regression

uncorrel_slope_intercept_pairs
boolean value to indicate interaction in the case of linear regression

random_slope_intercept_pairs
boolean value to indicate interaction in the case of linear regression

naaction
what action to be taken for the missing values, default is a missing value.

param2_to_be_estimated
parameter of interest for equation 2 in bivariate regression

covariates2
list of covariates - for equation 2 in bivariate regression

interaction2
boolean value to indicate interaction for equation 2 in bivariate regression

link
link function to be provided if not using the default link for each of the info_distribution

cluster_var
cluster variable if any

package_mixed_model
package to be used for mixed model ie nlme or lme4

Value

results the results of the regression analysis

Details

This function is the top in the layer of functions used for regression analysis Thus it contains many parameters to be passed on The required ones are parameter to be estimated, data that contains the observation, the method of regression to be used, the independent variable and the information for the distribution and method. if the data is given as a file name. it will load the data in that file Then it calls the appropriate functions depending on the regression method that specified. The methods that are considered : Survival analysis, linear regression, logistic regression,generalised linear model, linear multilevel or mixed model, and seemingly unrelated regression

Examples

# \donttest{
result <- get_parameter_estimated_regression(
  param_to_be_estimated = "Direction",
  data = ISLR::Smarket, method = "logistic", indep_var = "Lag1",
  info_get_method = NA, info_distribution = "binomial",
  covariates = c("Lag2", "Lag3"), interaction = FALSE,
  naaction = "na.omit", link = NA)

#> Waiting for profiling to be done...

#> Start:  AIC=1736.37
#> Direction ~ Lag2 + Lag3 + Lag1
#> 
#>        Df Deviance    AIC
#> - Lag3  1   1728.4 1734.4
#> - Lag2  1   1729.2 1735.2
#> <none>      1728.4 1736.4
#> - Lag1  1   1730.4 1736.4
#> 
#> Step:  AIC=1734.4
#> Direction ~ Lag2 + Lag1
#> 
#>        Df Deviance    AIC
#> - Lag2  1   1729.2 1733.2
#> <none>      1728.4 1734.4
#> - Lag1  1   1730.5 1734.5
#> + Lag3  1   1728.4 1736.4
#> 
#> Step:  AIC=1733.2
#> Direction ~ Lag1
#> 
#>        Df Deviance    AIC
#> - Lag1  1   1731.2 1733.2
#> <none>      1729.2 1733.2
#> + Lag2  1   1728.4 1734.4
#> + Lag3  1   1729.2 1735.2
#> 
#> Step:  AIC=1733.17
#> Direction ~ 1
#> 
#>        Df Deviance    AIC
#> <none>      1731.2 1733.2
#> + Lag1  1   1729.2 1733.2
#> + Lag2  1   1730.5 1734.5
#> + Lag3  1   1731.1 1735.1

  # }