Definition of Markov model and trace

Definition of Markov model and trace

markov_model(
  current_strategy,
  cycles,
  initial_state,
  discount = c(0, 0),
  parameter_values = NULL,
  half_cycle_correction = TRUE,
  state_cost_only_prevalent = FALSE,
  state_util_only_prevalent = FALSE,
  method = "half cycle correction",
  startup_cost = NULL,
  startup_util = NULL
)

Arguments

current_strategy
strategy object

cycles
no of cycles

initial_state
value of states initially

discount
rate of discount for costs and qalys

parameter_values
parameters for assigning health states and probabilities

half_cycle_correction
boolean to indicate half cycle correction

state_cost_only_prevalent
boolean parameter to indicate if the costs for state occupancy is only for those in the state excluding those that transitioned new. This is relevant when the transition cost is provided for eg. in a state with dialysis the cost of previous dialysis is different from the newly dialysis cases. Then the state_cost_only_prevalent should be TRUE

state_util_only_prevalent
boolean parameter to indicate if the utilities for state occupancy is only for those in the state excluding those that transitioned new.

method
what type of half cycle correction needed

startup_cost
cost of states initially

startup_util
utility of states initially if any

Value

Markov trace

Details

Use the strategy, cycles, initial state values creating the markov model and trace. As many probabilities /cost/utility value depend on age/time the evaluation and assignment happens during each cycle. At the heart it does a matrix multiplication using the previous row of the trace matrix and the columns of the transition matrix. Also checks for population loss, calculates cumulative costs and qalys (accounts for discounting and half cycle correction)

Examples

tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm <- populate_transition_matrix(2, tmat, c(0.5, 0.5, 0, 1))

#> [1] "The transition matrix as explained"
#>    transition number        probability name from from state to to state
#> 1:                 1 prob_Healthy_to_Healthy    1    Healthy  1  Healthy
#> 2:                 2    prob_Healthy_to_Dead    1    Healthy  2     Dead
#> 3:                 3    prob_Dead_to_Healthy    2       Dead  1  Healthy
#> 4:                 4       prob_Dead_to_Dead    2       Dead  2     Dead

a <- health_state("Healthy", 1, 1, 0, FALSE)
b <- health_state("Dead", 1, 0, 0, TRUE)
health_states <- combine_state(a, b)
this.strategy <- strategy(tm, health_states, "intervention")
markov_model(this.strategy, 10, c(1, 0))

#> $strategy
#> $name_strategy
#> [1] "intervention"
#> 
#> $transition_matrix
#> $trans_matrix
#>     1   2
#> 1 0.5 0.5
#> 2 0.0 1.0
#> 
#> $name_states
#> [1] 1 2
#> 
#> $no_states
#> [1] 2
#> 
#> attr(,"class")
#> [1] "transition_matrix"
#> 
#> $states
#> $states[[1]]
#> $name
#> [1] "Healthy"
#> 
#> $cost
#> [1] 1
#> 
#> $utility
#> [1] 1
#> 
#> $state_time
#> [1] 0
#> 
#> $absorb
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "health_state"
#> 
#> $states[[2]]
#> $name
#> [1] "Dead"
#> 
#> $cost
#> [1] 1
#> 
#> $utility
#> [1] 0
#> 
#> $state_time
#> [1] 0
#> 
#> $absorb
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "health_state"
#> 
#> 
#> $transition_cost
#> NULL
#> 
#> $transition_utility
#> NULL
#> 
#> attr(,"class")
#> [1] "strategy"
#> 
#> $method
#> [1] "half cycle correction"
#> 
#> $half_cycle_correction
#> [1] TRUE
#> 
#> $transition_matrix
#> $trans_matrix
#>        1   2
#> [1,] 0.5 0.5
#> [2,] 0.0 1.0
#> 
#> $name_states
#> [1] 1 2
#> 
#> $no_states
#> [1] 2
#> 
#> attr(,"class")
#> [1] "transition_matrix"
#> 
#> $param_matrix
#>       cycle
#>  [1,]     0
#>  [2,]     1
#>  [3,]     2
#>  [4,]     3
#>  [5,]     4
#>  [6,]     5
#>  [7,]     6
#>  [8,]     7
#>  [9,]     8
#> [10,]     9
#> [11,]    10
#> 
#> $list_param_values
#> NULL
#> 
#> $health_states
#> $health_states[[1]]
#> $name
#> [1] "Healthy"
#> 
#> $cost
#> [1] 1
#> 
#> $utility
#> [1] 1
#> 
#> $state_time
#> [1] 0
#> 
#> $absorb
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "health_state"
#> 
#> $health_states[[2]]
#> $name
#> [1] "Dead"
#> 
#> $cost
#> [1] 1
#> 
#> $utility
#> [1] 0
#> 
#> $state_time
#> [1] 0
#> 
#> $absorb
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "health_state"
#> 
#> 
#> $cycles
#> [1] 10
#> 
#> $initial_state
#> [1] 1 0
#> 
#> $discount
#> [1] 0 0
#> 
#> $trace_matrix
#>         Healthy      Dead Cycles
#> 0  1.0000000000 0.0000000      0
#> 1  0.5000000000 0.5000000      1
#> 2  0.2500000000 0.7500000      2
#> 3  0.1250000000 0.8750000      3
#> 4  0.0625000000 0.9375000      4
#> 5  0.0312500000 0.9687500      5
#> 6  0.0156250000 0.9843750      6
#> 7  0.0078125000 0.9921875      7
#> 8  0.0039062500 0.9960938      8
#> 9  0.0019531250 0.9980469      9
#> 10 0.0009765625 0.9990234     10
#> 
#> $cost_matrix
#>         Healthy      Dead Total discounted cost Cumulative cost
#> 0  1.0000000000 0.0000000                   0.5             0.5
#> 1  0.5000000000 0.5000000                   1.0             1.5
#> 2  0.2500000000 0.7500000                   1.0             2.5
#> 3  0.1250000000 0.8750000                   1.0             3.5
#> 4  0.0625000000 0.9375000                   1.0             4.5
#> 5  0.0312500000 0.9687500                   1.0             5.5
#> 6  0.0156250000 0.9843750                   1.0             6.5
#> 7  0.0078125000 0.9921875                   1.0             7.5
#> 8  0.0039062500 0.9960938                   1.0             8.5
#> 9  0.0019531250 0.9980469                   1.0             9.5
#> 10 0.0009765625 0.9990234                   0.5            10.0
#> 
#> $utility_matrix
#>         Healthy Dead Total discounted utility Cumulative utility
#> 0  1.0000000000    0             0.5000000000           0.500000
#> 1  0.5000000000    0             0.5000000000           1.000000
#> 2  0.2500000000    0             0.2500000000           1.250000
#> 3  0.1250000000    0             0.1250000000           1.375000
#> 4  0.0625000000    0             0.0625000000           1.437500
#> 5  0.0312500000    0             0.0312500000           1.468750
#> 6  0.0156250000    0             0.0156250000           1.484375
#> 7  0.0078125000    0             0.0078125000           1.492188
#> 8  0.0039062500    0             0.0039062500           1.496094
#> 9  0.0019531250    0             0.0019531250           1.498047
#> 10 0.0009765625    0             0.0004882812           1.498535
#> 
#> $startup_cost
#> NULL
#> 
#> $startup_util
#> NULL
#> 
#> attr(,"class")
#> [1] "markov_model"