Definition of strategy - or arm

Definition of strategy - or arm

strategy(trans_mat, states, name, trans_cost = NULL, trans_util = NULL)

Arguments

trans_mat
transition matrix

states
health states

name
name of the strategy

trans_cost
values of costs if these are attached to transitions

trans_util
values of utility if these are attached to transitions

Value

object strategy

Details

Defining strategy keeping all transition matrix, states and names together to use in defining Markov model

Examples

tmat <- rbind(c(1, 2), c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm <- populate_transition_matrix(2, tmat, c(0.5, 0.5, 0, 1))

#> [1] "The transition matrix as explained"
#>    transition number        probability name from from state to to state
#> 1:                 1 prob_Healthy_to_Healthy    1    Healthy  1  Healthy
#> 2:                 2    prob_Healthy_to_Dead    1    Healthy  2     Dead
#> 3:                 3    prob_Dead_to_Healthy    2       Dead  1  Healthy
#> 4:                 4       prob_Dead_to_Dead    2       Dead  2     Dead

a <- health_state("Healthy", 1, 1, 0, FALSE)
b <- health_state("Dead", 1, 0.5, 0, FALSE)
states <- combine_state(a, b)
strategy(tm, states, "intervention")

#> $name_strategy
#> [1] "intervention"
#> 
#> $transition_matrix
#> $trans_matrix
#>     1   2
#> 1 0.5 0.5
#> 2 0.0 1.0
#> 
#> $name_states
#> [1] 1 2
#> 
#> $no_states
#> [1] 2
#> 
#> attr(,"class")
#> [1] "transition_matrix"
#> 
#> $states
#> $states[[1]]
#> $name
#> [1] "Healthy"
#> 
#> $cost
#> [1] 1
#> 
#> $utility
#> [1] 1
#> 
#> $state_time
#> [1] 0
#> 
#> $absorb
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "health_state"
#> 
#> $states[[2]]
#> $name
#> [1] "Dead"
#> 
#> $cost
#> [1] 1
#> 
#> $utility
#> [1] 0.5
#> 
#> $state_time
#> [1] 0
#> 
#> $absorb
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "health_state"
#> 
#> 
#> $transition_cost
#> NULL
#> 
#> $transition_utility
#> NULL
#> 
#> attr(,"class")
#> [1] "strategy"