############################################################################ Get the parameter values using logistic regression

############################################################################ Get the parameter values using logistic regression

use_generalised_linear_model(
  param_to_be_estimated,
  dataset,
  indep_var,
  family,
  covariates,
  interaction,
  naaction,
  link = NA
)

Arguments

param_to_be_estimated
parameter of interest

dataset
data set to be provided

indep_var
the independent variable (column name in data file)

family
distribution name eg. for logistic regression -binomial

covariates
list of covariates-calculations to be done before passing

interaction
boolean value to indicate interaction in the case of linear regression

naaction
action to be taken with the missing values

link
link function if not the default for each family

Value

the results of the regression analysis

Details

This function returns the results and plots after doing linear regression Requires param to be estimated, dataset, independent variables and information on covariates, and interaction variables if there are Uses form_expression_glm to create the expression as per R standard for e.g glm(y ~ x ). Returns the fit result,s summary results as returned by summary(), confidence interval for fit coefficients (ci_coeff), variance covariance matrix, cholesky decomposition matrix, results from correlation test, plot of diagnostic tests and model fit assumptions, plot of model prediction diagnostic include AIC, R2, and BIC. The results of the prediction ie predicted values for fixed other variables will be returned in prediction matrix

Examples

# \donttest{
gm_result <- use_generalised_linear_model(
  param_to_be_estimated = "Direction",
  dataset = ISLR::Smarket, indep_var = "Lag1", family = "binomial",
  covariates = c("Lag2", "Lag3"),
  interaction = FALSE, naaction = "na.omit", link = NA)

#> Waiting for profiling to be done...

#> Start:  AIC=1736.37
#> Direction ~ Lag2 + Lag3 + Lag1
#> 
#>        Df Deviance    AIC
#> - Lag3  1   1728.4 1734.4
#> - Lag2  1   1729.2 1735.2
#> <none>      1728.4 1736.4
#> - Lag1  1   1730.4 1736.4
#> 
#> Step:  AIC=1734.4
#> Direction ~ Lag2 + Lag1
#> 
#>        Df Deviance    AIC
#> - Lag2  1   1729.2 1733.2
#> <none>      1728.4 1734.4
#> - Lag1  1   1730.5 1734.5
#> + Lag3  1   1728.4 1736.4
#> 
#> Step:  AIC=1733.2
#> Direction ~ Lag1
#> 
#>        Df Deviance    AIC
#> - Lag1  1   1731.2 1733.2
#> <none>      1729.2 1733.2
#> + Lag2  1   1728.4 1734.4
#> + Lag3  1   1729.2 1735.2
#> 
#> Step:  AIC=1733.17
#> Direction ~ 1
#> 
#>        Df Deviance    AIC
#> <none>      1731.2 1733.2
#> + Lag1  1   1729.2 1733.2
#> + Lag2  1   1730.5 1734.5
#> + Lag3  1   1731.1 1735.1

# }