########################################################################### Get the parameter values using the linear regression

########################################################################### Get the parameter values using the linear regression

use_linear_regression(
  param_to_be_estimated,
  dataset,
  indep_var,
  covariates,
  interaction
)

Arguments

param_to_be_estimated
parameter of interest

dataset
data set to be provided

indep_var
the independent variable (column name in data file)

covariates
list of covariates-calculations to be done before passing

interaction
boolean value to indicate interaction in the case of linear regression, false by default

Value

the results of the regression analysis

Details

This function returns the results and plots after doing linear regression Requires param to be estimated, dataset, independent variables and information on covariates, and interaction variables if there are Uses form_expression_lm to create the expression as per R standard for e.g lm(y ~ x ). Returns the fit result,s summary results as returned by summary(), confidence interval for fit coefficients (ci_coeff), variance covariance matrix, cholesky decomposition matrix, Results from correlation test, plot of diagnostic tests and model fit assumptions, plot of model prediction diagnostic include AIC, R2, and BIC. The results of the prediction ie predicted values when each of covariate is fixed will be returned in prediction matrix predicted values will provide the mean value of param_to_to_estimated as calculated by the linear regression formula. ref:https://www.statmethods.net/stats/regression.html

Examples

# \donttest{
results_lm <- use_linear_regression("dist",
  dataset = cars,
  indep_var = "speed", covariates = NA, interaction = FALSE)

#> 
#> Call:
#> lm(formula = dist ~ speed, data = dataset)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -29.069  -9.525  -2.272   9.215  43.201 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) -17.5791     6.7584  -2.601   0.0123 *  
#> speed         3.9324     0.4155   9.464 1.49e-12 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 15.38 on 48 degrees of freedom
#> Multiple R-squared:  0.6511,	Adjusted R-squared:  0.6438 
#> F-statistic: 89.57 on 1 and 48 DF,  p-value: 1.49e-12
#> 
#> 
#> ASSESSMENT OF THE LINEAR MODEL ASSUMPTIONS
#> USING THE GLOBAL TEST ON 4 DEGREES-OF-FREEDOM:
#> Level of Significance =  0.05 
#> 
#> Call:
#>  gvlma::gvlma(x = fit) 
#> 
#>                     Value  p-value                   Decision
#> Global Stat        15.801 0.003298 Assumptions NOT satisfied!
#> Skewness            6.528 0.010621 Assumptions NOT satisfied!
#> Kurtosis            1.661 0.197449    Assumptions acceptable.
#> Link Function       2.329 0.126998    Assumptions acceptable.
#> Heteroscedasticity  5.283 0.021530 Assumptions NOT satisfied!

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

# }
# \donttest{
library(car)

#> Loading required package: carData

results_lm <- use_linear_regression("mpg",
  dataset = mtcars,
  indep_var = "disp", covariates = c("hp", "wt", "drat"),
  interaction = FALSE)

#> Start:  AIC=65.77
#> mpg ~ hp + wt + drat + disp
#> 
#>        Df Sum of Sq    RSS    AIC
#> - disp  1     0.844 183.68 63.919
#> <none>              182.84 65.772
#> - drat  1    12.153 194.99 65.831
#> - hp    1    60.916 243.75 72.974
#> - wt    1    70.508 253.35 74.209
#> 
#> Step:  AIC=63.92
#> mpg ~ hp + wt + drat
#> 
#>        Df Sum of Sq    RSS    AIC
#> - drat  1    11.366 195.05 63.840
#> <none>              183.68 63.919
#> + disp  1     0.844 182.84 65.772
#> - hp    1    85.559 269.24 74.156
#> - wt    1   107.771 291.45 76.693
#> 
#> Step:  AIC=63.84
#> mpg ~ hp + wt
#> 
#>        Df Sum of Sq    RSS    AIC
#> <none>              195.05 63.840
#> + drat  1    11.366 183.68 63.919
#> + disp  1     0.057 194.99 65.831
#> - hp    1    83.274 278.32 73.217
#> - wt    1   252.627 447.67 88.427
#> 
#> Call:
#> lm(formula = mpg ~ hp + wt + drat + disp, data = dataset)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.5077 -1.9052 -0.5057  0.9821  5.6883 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 29.148738   6.293588   4.631  8.2e-05 ***
#> hp          -0.034784   0.011597  -2.999  0.00576 ** 
#> wt          -3.479668   1.078371  -3.227  0.00327 ** 
#> drat         1.768049   1.319779   1.340  0.19153    
#> disp         0.003815   0.010805   0.353  0.72675    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 2.602 on 27 degrees of freedom
#> Multiple R-squared:  0.8376,	Adjusted R-squared:  0.8136 
#> F-statistic: 34.82 on 4 and 27 DF,  p-value: 2.704e-10
#> 
#> 
#> ASSESSMENT OF THE LINEAR MODEL ASSUMPTIONS
#> USING THE GLOBAL TEST ON 4 DEGREES-OF-FREEDOM:
#> Level of Significance =  0.05 
#> 
#> Call:
#>  gvlma::gvlma(x = fit) 
#> 
#>                       Value  p-value                   Decision
#> Global Stat        13.93816 0.007495 Assumptions NOT satisfied!
#> Skewness            4.31310 0.037820 Assumptions NOT satisfied!
#> Kurtosis            0.01378 0.906542    Assumptions acceptable.
#> Link Function       8.71658 0.003153 Assumptions NOT satisfied!
#> Heteroscedasticity  0.89470 0.344207    Assumptions acceptable.

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

#> `geom_smooth()` using formula 'y ~ x'

# }