Function to estimate the cost of tablets taken (from IPD)

Function to estimate the cost of tablets taken (from IPD)

value_Shows_IPD(ind_part_data, shows_related_words, shows_nrcode)

Arguments

ind_part_data
a dataframe containing IPD

shows_related_words
a dataframe containing IPD

shows_nrcode
non response code for ADL, default is NA

Value

sum of scores, if success -1, if failure

Examples

datafile <- system.file("extdata", "trial_data.csv", package = "packDAMipd")
trial_data <- load_trial_data(datafile)
value_Shows_IPD(trial_data, "qsy", NA)

#>    pat_TrialArm SEX age                    Drug Timepoint       Start.Dose
#> 1             N   f  44       ORAMORPH 10MG/5ML         2       300ML/WEEK
#> 2             N   f  30          matrifen patch         1           200mcg
#> 3             N   M  21                 Zomorph         2          40mg bd
#> 4             N   M  43             Tramadol MR         2         200mg BD
#> 5             Y   F  37                 Zomorph         2   30mg bd (60mg)
#> 6             Y   F  73               Oxycontin         1   30mg bd (60mg)
#> 7             N   F  46 fentanyl 50 and 25 ptch         1          75mcg/h
#> 8             N   F  53         oxynorm 10mg/ml         1  6ml/6 doses/day
#> 9             N   F  84            Tramadol MR          1         200mg BD
#> 10            N   F  87                 Zomorph         1          60mg BD
#> 11            N   F  65          matrifen patch         1            62mcg
#> 12            Y   M  75          fentanyl patch         2           150mcg
#> 13            N   F  35                 Zomorph         2          40mg bd
#> 14            N   F  53  Buprenorphine patches          2       20mcg/hour
#> 15            N   F  82                   Marol         2         200mg bd
#> 16            N   F  78                     MST         1            70 bd
#> 17            N   F  64       ORAMORPH 10MG/5ML         2     1200ml/month
#> 18            N   F  40                 zomorph         2          70mg bd
#> 19            Y   F  52           mezolar patch         2          50mcg/h
#> 20            N   M  65       Buprenorphine s/l         2     200mcg 8/day
#> 21            N   M  64           BUPRENORPHINE         1          70MCG/H
#> 22            Y   F  20           BUPRENORPHINE         1          35MCG/H
#> 23            Y   M  84      OXYCODONE/ORAMORPH         2          40MG BD
#> 24            N   F  62                 ZOMORPH         1          60MG BD
#> 25            N   F  37           BUPRENORPHINE         1           5MCG/H
#> 26            N   F  48             TRAMADOL MR         1         200MG BD
#> 27            N   M  52                 ZOMORPH         1          10MG BD
#> 28            Y   F  89    TARGINACT / ORAMORPH         1    40/20  / 30MG
#> 29            N   M  87                FENTANYL         2          12MGC/H
#> 30            N   F  29               OXYCONTIN         1         90MG/DAY
#> 31            Y   M  83          ORAMORPH / MST         1 200MG / 100MG BD
#>    start.MME Final.Dose end.MME Reduction.months qol_Mobility qol_SelfCare
#> 1         86          0       0                3            1            3
#> 2        480      50mcg     120               13            3            5
#> 3         80          0       0                2            4            1
#> 4         40                  0                3            3            5
#> 5         60                  0                2            1            3
#> 6         91                  0                2            3            1
#> 7        180                  0                3            5            1
#> 8        545                  0                5            5            1
#> 9         40                  0                8            2            2
#> 10       120                  0                4            3            2
#> 11       150                  0                4            4            2
#> 12       360      25mcg      60                5            2            1
#> 13        80                  0                6            1            5
#> 14        48    20mcg/h      48                0            5            4
#> 15       400                  0                2            5            5
#> 16       140    10gm bd      20               10            5            2
#> 17        80                  0                6            5            1
#> 18       140                  0                8            2            2
#> 19       120                  0                6            1            3
#> 20       120    1600mcg     120        restarted            3            4
#> 21       170          0       0                3            1            5
#> 22        85          0       0                4            3            4
#> 23       160          0       0                4            1            3
#> 24       120          0       0                8            3            2
#> 25        12          0       0                1            5            2
#> 26        40          0       0                3            1            4
#> 27        20          0       0                2            4            5
#> 28       150          0       0                9            1            2
#> 29        30          0       0                2            4            4
#> 30       140          0       0                5            3            2
#> 31       400          0       0                9            4            2
#>    qol_UsualActivity qol_Pain qol_Anxiety tpi_PainDayToDayActivities
#> 1                  2        5           1                          1
#> 2                  5        2           3                          4
#> 3                  1        2           3                          1
#> 4                  4        5           1                          2
#> 5                  4        5           1                          2
#> 6                  2        3           2                          1
#> 7                  3        1           2                          3
#> 8                  4        5           1                          1
#> 9                  5        4           5                          1
#> 10                 3        2           4                          5
#> 11                 4        1           2                          4
#> 12                 5        2           4                          1
#> 13                 3        1           5                          5
#> 14                 3        3           5                          2
#> 15                 4        3           3                          5
#> 16                 1        2           5                          5
#> 17                 3        4           1                          3
#> 18                 3        2           2                          2
#> 19                 5        5           3                          5
#> 20                 1        5           5                          3
#> 21                 3        1           3                          5
#> 22                 2        5           4                          2
#> 23                 5        3           4                          3
#> 24                 5        3           3                          2
#> 25                 3        3           5                          1
#> 26                 4        2           4                          3
#> 27                 5        2           4                          2
#> 28                 2        1           5                          1
#> 29                 3        4           2                          2
#> 30                 5        3           5                          3
#> 31                 1        2           5                          3
#>    tpi_PainWorkAroundHome tpi_PainParticipateSocial tpi_PainHouseholdChores
#> 1                       1                         4                       4
#> 2                       2                         4                       1
#> 3                       1                         5                       5
#> 4                       2                         5                       1
#> 5                       1                         2                       1
#> 6                       1                         5                       2
#> 7                       4                         4                       1
#> 8                       4                         1                       1
#> 9                       2                         4                       3
#> 10                      3                         4                       1
#> 11                      1                         2                       5
#> 12                      3                         2                       4
#> 13                      2                         3                       1
#> 14                      5                         1                       4
#> 15                      2                         4                       3
#> 16                      5                         1                       1
#> 17                      1                         5                       5
#> 18                      3                         5                       1
#> 19                      4                         1                       2
#> 20                      4                         2                       4
#> 21                      4                         1                       2
#> 22                      4                         3                       5
#> 23                      2                         4                       4
#> 24                      2                         4                       1
#> 25                      3                         2                       1
#> 26                      4                         4                       5
#> 27                      1                         5                       5
#> 28                      2                         3                       1
#> 29                      1                         2                       1
#> 30                      4                         5                       5
#> 31                      5                         5                       3
#>    tpi_PainUsuallyDoForFun tpi_PainEnjoySocial tpi_PainEnjoyLife
#> 1                        4                   3                 1
#> 2                        5                   4                 1
#> 3                        1                   5                 1
#> 4                        4                   2                 1
#> 5                        5                   4                 5
#> 6                        4                   1                 5
#> 7                        2                   5                 4
#> 8                        2                   2                 3
#> 9                        2                   4                 2
#> 10                       5                   4                 1
#> 11                       2                   2                 5
#> 12                       5                   4                 4
#> 13                       5                   4                 2
#> 14                       4                   2                 5
#> 15                       1                   5                 1
#> 16                       4                   3                 4
#> 17                       5                   4                 2
#> 18                       3                   2                 2
#> 19                       4                   1                 1
#> 20                       3                   2                 1
#> 21                       5                   2                 2
#> 22                       5                   2                 3
#> 23                       2                   5                 3
#> 24                       4                   3                 5
#> 25                       1                   5                 1
#> 26                       4                   3                 3
#> 27                       4                   5                 5
#> 28                       3                   4                 4
#> 29                       4                   5                 2
#> 30                       3                   3                 5
#> 31                       2                   4                 4
#>    tpi_PainFamilyLife qsy_SympFeelingSick qsy_SympStomachCramps
#> 1                   4                   4                     2
#> 2                   4                   1                     4
#> 3                   1                   2                     2
#> 4                   2                   2                     1
#> 5                   3                   2                     3
#> 6                   2                   3                     1
#> 7                   1                   1                     3
#> 8                   1                   1                     4
#> 9                   3                   1                     4
#> 10                  1                   4                     1
#> 11                  4                   4                     3
#> 12                  5                   1                     2
#> 13                  5                   1                     3
#> 14                  1                   4                     2
#> 15                  2                   1                     2
#> 16                  1                   3                     4
#> 17                  1                   2                     4
#> 18                  5                   4                     4
#> 19                  3                   3                     2
#> 20                  4                   1                     1
#> 21                  1                   2                     1
#> 22                  4                   2                     3
#> 23                  2                   1                     3
#> 24                  1                   3                     2
#> 25                  5                   1                     3
#> 26                  3                   2                     1
#> 27                  4                   3                     3
#> 28                  2                   4                     4
#> 29                  5                   4                     2
#> 30                  2                   4                     1
#> 31                  1                   3                     1
#>    qsy_SympMuscleSpasms qsy_SympFeelColdness qsy_SympHeartPounding
#> 1                     3                    4                     1
#> 2                     3                    3                     3
#> 3                     4                    3                     2
#> 4                     2                    4                     2
#> 5                     3                    4                     3
#> 6                     2                    4                     3
#> 7                     2                    2                     3
#> 8                     1                    2                     3
#> 9                     2                    1                     1
#> 10                    2                    3                     1
#> 11                    2                    2                     3
#> 12                    2                    1                     2
#> 13                    2                    3                     2
#> 14                    4                    1                     1
#> 15                    4                    3                     3
#> 16                    4                    3                     3
#> 17                    2                    2                     1
#> 18                    2                    4                     1
#> 19                    3                    3                     3
#> 20                    4                    3                     3
#> 21                    2                    4                     4
#> 22                    4                    2                     1
#> 23                    1                    4                     1
#> 24                    3                    1                     1
#> 25                    3                    3                     4
#> 26                    2                    1                     3
#> 27                    4                    2                     4
#> 28                    4                    3                     2
#> 29                    2                    4                     2
#> 30                    4                    4                     4
#> 31                    4                    4                     4
#>    qsy_SympMuscularTension qsy_SympAches qsy_SympYawning qsy_SympRunnyEyes
#> 1                        4             3               2                 2
#> 2                        4             3               4                 2
#> 3                        3             2               3                 1
#> 4                        1             3               1                 3
#> 5                        1             1               4                 3
#> 6                        3             2               3                 2
#> 7                        3             2               1                 1
#> 8                        3             1               3                 3
#> 9                        1             1               1                 4
#> 10                       4             4               4                 4
#> 11                       1             4               2                 4
#> 12                       4             4               2                 2
#> 13                       3             3               2                 1
#> 14                       4             3               3                 3
#> 15                       3             1               2                 2
#> 16                       4             4               3                 3
#> 17                       2             3               4                 3
#> 18                       1             3               4                 1
#> 19                       3             2               3                 1
#> 20                       3             2               3                 2
#> 21                       4             4               2                 4
#> 22                       1             3               4                 2
#> 23                       4             1               2                 3
#> 24                       2             4               4                 3
#> 25                       2             1               2                 2
#> 26                       3             2               2                 3
#> 27                       1             3               4                 4
#> 28                       1             1               2                 3
#> 29                       3             4               4                 1
#> 30                       2             3               2                 3
#> 31                       4             2               2                 4
#>    qsy_SympInsomnia ShOWSscore
#> 1                 1         16
#> 2                 1         18
#> 3                 3         15
#> 4                 1         10
#> 5                 3         17
#> 6                 4         17
#> 7                 3         11
#> 8                 1         12
#> 9                 1          7
#> 10                1         18
#> 11                3         18
#> 12                4         14
#> 13                4         14
#> 14                1         16
#> 15                4         15
#> 16                3         24
#> 17                2         15
#> 18                1         15
#> 19                4         17
#> 20                1         13
#> 21                4         21
#> 22                4         16
#> 23                2         12
#> 24                1         14
#> 25                2         13
#> 26                1         10
#> 27                2         20
#> 28                4         18
#> 29                2         18
#> 30                2         19
#> 31                3         21